requirements.txt
firebase_admin==6.5.0
Flask==3.0.3
fpdf==1.7.2
# Pillow==9.0.1
Pillow==10.3.0
pyTelegramBotAPI==4.19.2
Requests==2.32.3
